Give the search shapes their own package: search/strategy/
Mirrors search/policy/ so the layer taxonomy is the directory tree:
policy/ answers the questions one loop asks; strategy/ composes loops.
base.py (SearchStrategy ABC), greedy.py (GreedyStrategy + its retry
helpers, moved out of policy/greedy.py — the decide callback and pricing
stay there, since that is the policy), two_level.py (moved wholesale).
