+++ emmy/commands/tune.py
@@ -331,7 +331,7 @@ def _tune_one(
     import time
 
     from emmy.commands.tune_progress import TuneProgress
-    from emmy.compiler.pipeline.search.two_level import TwoLevelStrategy
+    from emmy.compiler.pipeline.search.strategy import TwoLevelStrategy
 
     graph, _, bench_bundle = load_or_trace(args)
     if dump:
@@ -473,7 +473,7 @@ def _tune_working_multi(args, targets, document, *, backends, db, ctx, run_id) -
     """
     from emmy.commands.tune_progress import TuneProgress
     from emmy.compiler.pipeline.search.prior import load_prior
-    from emmy.compiler.pipeline.search.two_level import TwoLevelStrategy
+    from emmy.compiler.pipeline.search.strategy import TwoLevelStrategy
 
     prepared = []
     temp_dumps: list[Path] = []
@@ -652,7 +652,7 @@ def handle_tune(args):
     # One session id per CLI invocation (a golden sweep = one collection session) —
     # stamped on every node row this run writes, so cross-run keep-min drift in the
     # node store is traceable to its tune session.
-    from emmy.compiler.pipeline.search.two_level import _mint_run_id
+    from emmy.compiler.pipeline.search.strategy.two_level import _mint_run_id
 
     run_id = _mint_run_id()
