+++ emmy/compiler/pipeline/search/strategy/base.py
